# PRN232.LMS.Services/Models/Requests/StudentRequest.cs
namespace PRN232.LMS.Services.Models.Requests;

public class CreateStudentRequest
{
    public string FullName { get; set; } = null!;
    public string Email { get; set; } = null!;
    public DateTime DateOfBirth { get; set; }
}

public class UpdateStudentRequest
{
    public string FullName { get; set; } = null!;
    public string Email { get; set; } = null!;
    public DateTime DateOfBirth { get; set; }
}